Code cleanup of non-gui tests.
I found a few items that are low priority and ignorable.

This item could be fixed by creating a function, but would require
more lines at very little improvement of readability imo.  May fix or
solve in a different way but for now it stays.
./test_lib.py:52:9: E731 do not assign a lambda expression, use a def

This item is an artifact of appeding the python module path before
importing a local module. This should be igmored.
./test_yieldfrom.py:9:1: E402 module level import not at top of file